+++ R/pooledSPsurv.R
@@ -13,8 +13,6 @@
 #' @param m limit on steps in the slice sampling. A vector of values for beta, gamma, rho.
 #' @param ini.beta ...
 #' @param ini.gamma ...
-#' @param rho ...
-#' @param lambda ...
 #' @param ini.W ...
 #' @param ini.V ...
 #' @param form type of parametric model (Weibull, Exponential, or Log-Logistic).
@@ -33,8 +31,6 @@
 #' \item{C}{vector of `C'.}
 #' \item{ini.beta}{...}
 #' \item{ini.gamma}{...}
-#' \item{rho}{...}
-#' \item{lambda}{...}
 #' \item{ini.W}{...}
 #' \item{ini.V}{...}
 #' \item{form}{character, type of distribution.}
@@ -51,7 +47,7 @@
 #' model <-
 #'     pooledSPsurv(
 #'         duration = duration ~ fhcompor1 + lgdpl + comprehensive + victory +
-#'             instabl + intensityln + ethfrac + unpko,
+#'                    instabl + intensityln + ethfrac + unpko,
 #'         immune   = cured ~ fhcompor1 + lgdpl + victory,
 #'         Y0       = 't.0',
 #'         LY       = 'lastyear',
@@ -88,8 +84,6 @@ pooledSPsurv <- function(duration,
                   m = 10,
                   ini.beta =  0,
                   ini.gamma = 0,
-                  rho = 1,
-                  lambda = 1,
                   ini.W = 0,
                   ini.V= 0,
                   form = c('Weibull', 'exponential', 'loglog'))
@@ -99,21 +93,20 @@ pooledSPsurv <- function(duration,
     model <- 'SPsurv'
     r   <- formcall(duration = duration, immune = immune, data = data, Y0 = Y0,
                     LY = LY, N = N, burn = burn, thin = thin, w = w,
-                    m = m, ini.beta = ini.beta, ini.gamma = ini.gamma, rho = rho,
-                    lambda = lambda, ini.W = ini.W, ini.V = ini.V,
+                    m = m, ini.beta = ini.beta, ini.gamma = ini.gamma, ini.W = ini.W, ini.V = ini.V,
                     form = dis, model = model)
 
     if(form == 'loglog') {
         results <- mcmcSPlog(Y = r$Y, Y0 = r$Y0, C = r$C, LY = r$LY, X = r$X, Z = r$Z,
                           N = r$N, burn = r$burn, thin = r$thin, w  = r$w, m  = r$m,
                           ini.beta = r$ini.beta, ini.gamma = r$ini.gamma,
-                          rho = r$rho, lambda = r$lambda, ini.W = r$ini.W, ini.V = r$ini.V,
+                          ini.W = r$ini.W, ini.V = r$ini.V,
                           form = r$form)
     } else {
         results <- mcmcSP(Y = r$Y, Y0 = r$Y0, C = r$C, LY = r$LY, X = r$X, Z = r$Z,
                       N = r$N, burn = r$burn, thin = r$thin, w  = r$w, m  = r$m,
                       ini.beta = r$ini.beta, ini.gamma = r$ini.gamma,
-                      rho = r$rho, lambda = r$lambda, ini.W = r$ini.W, ini.V = r$ini.V,
+                      ini.W = r$ini.W, ini.V = r$ini.V,
                       form = r$form)
     }